refactor: separate deep research event models
